refactor: relocate scorers into per-feature directories (#108 phase 8)
