+++ src/lib/ai/features/game-plan/pivot-explanation.test.ts
