+++ src/lib/ai/features/game-plan/state-awareness.test.ts
