+++ src/lib/ai/features/item-rec/gold-aware-recommendations.test.ts
