+++ src/lib/ai/features/item-rec/item-awareness.test.ts
