+++ src/lib/ai/features/voice-query/gold-awareness.test.ts
